server/requirements.txt
Flask==3.1.2
requests==2.32.5
gunicorn==21.2.0
python-dotenv
langchain_groq
langchain_core